--- src/App.css
@@ -1,120 +1,0 @@
-#root {
-  max-width: 1280px;
-  margin: 0 auto;
-  padding: 2rem;
-  text-align: center;
-}
-
-.game {
-  display: flex;
-  flex-direction: column;
-}
-
-.title {
-  font-size: 60px;
-  font-weight: bold;
-}
-
-.header {
-  display: flex;
-  flex-direction: left;
-  justify-content: space-evenly;
-}
-
-.game {
-  width: 500px;
-  margin: 0 auto;
-}
-
-.origin-link {
-  width: 80px;
-  height: 80px;
-}
-
-.board {
-  display: grid;
-  grid-template-columns: repeat(4, 1fr);
-  grid-gap: 15px;
-  background-color: #7cbff0;
-  border-radius: 6px;
-  padding: 15px;
-}
-
-.tile {
-  width: 100px;
-  height: 100px;
-  background-color: #eefcff;
-  border-radius: 3px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  font-size: 45px;
-  font-weight: bold;
-  color: #776e65;
-}
-
-.tile-2 {
-  background-color: #aae7ff;
-}
-.tile-4 {
-  background-color: #bbe5ff;
-}
-.tile-8 {
-  background-color: #ccddff;
-}
-.tile-16 {
-  background-color: #ddccff;
-}
-.tile-32 {
-  background-color: #eebbff;
-}
-.tile-64 {
-  background-color: #ffccff;
-}
-.tile-128 {
-  background-color: #ddccff;
-}
-
-.game-header {
-  display: flex;
-  flex-direction: left;
-}
-
-.game-message {
-  width: 20%;
-  height: 75px;
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  background-color: #ffffff;
-  color: #2c2825;
-  text-align: center;
-  font-size: 48px;
-  font-weight: bold;
-  border-radius: 20px;
-  border: 4px solid #49abf1;
-}
-
-.score-container {
-  display: block;
-  margin: 15px auto;
-  padding: 10px 30px;
-  font-size: 20px;
-  background-color: #7cbff0;
-  color: #f9f6f2;
-  border: none;
-  border-radius: 7px;
-}
-
-button {
-  display: block;
-  margin: 15px auto;
-  padding: 10px 40px;
-  font-size: 20px;
-  background-color: #7cbff0;
-  color: #f9f6f2;
-  border: none;
-  border-radius: 7px;
-  cursor: pointer;
-}
